backend/pyproject.toml
[tool.poetry]
name = "headliner"
version = "0.1.0"
description = ""
authors = ["4got10king <Aldavletshina77654@yandex.ru>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.9"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.6.0"
pydantic-settings = "^2.8.1"
aio-pika = "^9.4.0"
pillow = "^10.2.0"
aiohttp = "^3.9.0"
aiosqlite = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
